Fixed termination conditions in the client code.
Modified the termination conditions in two ways.

First, the connection is not closed and the loop for ferrying
information does not end until the server closes the connection or there
is an error reading from the server.

Second, if the file descriptor for input closes, shutdown() is used to
signal to the server that we are done sending information.  This allows
the client to continue reading data from the server until the server is
finished.